backend/requirements.txt
dj-database-url==2.1.0

Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-extensions>=3.2.0
ipython>=8.0.0
Pillow>=10.0.0
python-decouple>=3.8
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
black>=23.0.0
# Production dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

